+++ packages/vlossom/src/components/vs-block/stories/VsBlock.stories.ts
@@ -1,6 +1,5 @@
-import { colorScheme, getColorSchemeTemplate } from '@/storybook/args';
+import { chromaticParameters, colorScheme, getColorSchemeTemplate } from '@/storybook';
 import VsBlock from './../VsBlock.vue';
-import { chromaticParameters } from '@/storybook/parameters';
 
 import type { Meta, StoryObj } from '@storybook/vue3';
