+++ packages/vlossom/src/components/vs-button/stories/VsButton.stories.ts
@@ -1,5 +1,4 @@
-import { colorScheme, getColorSchemeTemplate } from '@/storybook/args';
-import { chromaticParameters } from '@/storybook/parameters';
+import { chromaticParameters, colorScheme, getColorSchemeTemplate } from '@/storybook';
 import VsButton from './../VsButton.vue';
 
 import type { Meta, StoryObj } from '@storybook/vue3';
