+++ packages/vlossom/src/components/vs-checkbox-set/stories/VsCheckboxSet.stories.ts
@@ -1,8 +1,7 @@
-import { colorScheme, getMetaArguments } from '@/storybook/args';
-import VsCheckboxSet from '../VsCheckboxSet.vue';
+import { chromaticParameters, colorScheme, getMetaArguments } from '@/storybook';
 import { UIState } from '@/declaration';
 import VsContainer from '@/components/vs-container/VsContainer.vue';
-import { chromaticParameters } from '@/storybook/parameters';
+import VsCheckboxSet from '../VsCheckboxSet.vue';
 
 import type { Meta, StoryObj } from '@storybook/vue3';
