+++ packages/vlossom/src/components/vs-notice/stories/VsNotice.stories.ts
@@ -1,6 +1,5 @@
-import { colorScheme, getColorSchemeTemplate } from '@/storybook/args';
+import { chromaticParameters, colorScheme, getColorSchemeTemplate } from '@/storybook';
 import VsNotice from '../VsNotice.vue';
-import { chromaticParameters } from '@/storybook/parameters';
 
 import type { Meta, StoryObj } from '@storybook/vue3';
