+++ packages/vlossom/src/components/vs-text-wrap/stories/VsTextWrap.stories.ts
@@ -1,7 +1,7 @@
-import { colorScheme, align, placement, getColorSchemeTemplate } from '@/storybook/args';
+import { chromaticParameters, colorScheme, align, placement, getColorSchemeTemplate } from '@/storybook';
+import { PLACEMENTS, ALIGNS } from '@/declaration';
 import VsTextWrap from './../VsTextWrap.vue';
-import { chromaticParameters } from '@/storybook/parameters';
-import { PLACEMENTS, ALIGNS } from '@/declaration/constants';
+
 import type { Meta, StoryObj } from '@storybook/vue3';
 
 const meta: Meta<typeof VsTextWrap> = {
